Create Open Roles
This repository is a way of seeing open roles at Vauban, on Github. We mirror these roles and their descriptions to several other platforms. All open roles as in the base level of this repository. 

There's a ./inactive directory that contains roles we've already hired people into or are no longer active. We keep it there to give potential candidates context on the roles we've hired for in the past and how we describe them. 

If you have any questions about Vauban, our open roles or how we work as a team please contact careers@vauban.io. We'd love to talk with you.